ai/requirements.txt
fastapi
uvicorn[standard]
python-multipart

PyMuPDF
python-docx
easyocr
pytesseract
Pillow
pdfplumber

langdetect
spacy
sentence-transformers
faiss-cpu

transformers
torch